+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Change Log
 ### Features
 * Cria estrutura para salvar requisições e respostas das integrações. (PL-2306)
 
+## Stage
+
+### Features
+* Cria funcionalidade de testes do aferidor (PL-2343)
+
 <a name="v1.0.3"></a>
 ## Version [v1.0.3](https://github.com/luizfp/PrologWebService/compare/v1.0.2...v1.0.3) (release-date) [unreleased]